clone a slightly larger repo

+++ src/commands/checkout.js
@@ -131,26 +131,22 @@ export async function _checkout({
     let count = 0
     const total = ops.length
     await GitIndexManager.acquire({ fs, gitdir, cache }, async function(index) {
-      await Promise.all(
-        ops
-          .filter(
-            ([method]) => method === 'delete' || method === 'delete-index'
-          )
-          .map(async function([method, fullpath]) {
-            const filepath = `${dir}/${fullpath}`
-            if (method === 'delete') {
-              await fs.rm(filepath)
-            }
-            index.delete({ filepath: fullpath })
-            if (onProgress) {
-              await onProgress({
-                phase: 'Updating workdir',
-                loaded: ++count,
-                total,
-              })
-            }
-          })
-      )
+      for (const [method, fullpath] of ops) {
+        if (method === 'delete' || method === 'delete-index') {
+          const filepath = `${dir}/${fullpath}`
+          if (method === 'delete') {
+            await fs.rm(filepath)
+          }
+          index.delete({ filepath: fullpath })
+          if (onProgress) {
+            await onProgress({
+              phase: 'Updating workdir',
+              loaded: ++count,
+              total,
+            })
+          }
+        }
+      }
     })
 
     // Note: this is cannot be done naively in parallel
@@ -183,87 +179,82 @@ export async function _checkout({
       }
     })
 
-    await Promise.all(
-      ops
-        .filter(([method]) => method === 'mkdir' || method === 'mkdir-index')
-        .map(async function([_, fullpath]) {
-          const filepath = `${dir}/${fullpath}`
-          await fs.mkdir(filepath)
-          if (onProgress) {
-            await onProgress({
-              phase: 'Updating workdir',
-              loaded: ++count,
-              total,
-            })
-          }
-        })
-    )
+    for (const [method, fullpath] of ops) {
+      if (method === 'mkdir' || method === 'mkdir-index') {
+        const filepath = `${dir}/${fullpath}`
+        await fs.mkdir(filepath)
+        if (onProgress) {
+          await onProgress({
+            phase: 'Updating workdir',
+            loaded: ++count,
+            total,
+          })
+        }
+      }
+    }
 
     await GitIndexManager.acquire({ fs, gitdir, cache }, async function(index) {
-      await Promise.all(
-        ops
-          .filter(
-            ([method]) =>
-              method === 'create' ||
-              method === 'create-index' ||
-              method === 'update' ||
-              method === 'mkdir-index'
-          )
-          .map(async function([method, fullpath, oid, mode, chmod]) {
-            const filepath = `${dir}/${fullpath}`
-            try {
-              if (method !== 'create-index' && method !== 'mkdir-index') {
-                const { object } = await readObject({ fs, gitdir, oid })
-                if (chmod) {
-                  // Note: the mode option of fs.write only works when creating files,
-                  // not updating them. Since the `fs` plugin doesn't expose `chmod` this
-                  // is our only option.
-                  await fs.rm(filepath)
-                }
-                if (mode === 0o100644) {
-                  // regular file
-                  await fs.write(filepath, object)
-                } else if (mode === 0o100755) {
-                  // executable file
-                  await fs.write(filepath, object, { mode: 0o777 })
-                } else if (mode === 0o120000) {
-                  // symlink
-                  await fs.writelink(filepath, object)
-                } else {
-                  throw new InternalError(
-                    `Invalid mode 0o${mode.toString(8)} detected in blob ${oid}`
-                  )
-                }
+      for (const [method, fullpath, oid, mode, chmod] of ops) {
+        if (
+          method === 'create' ||
+          method === 'create-index' ||
+          method === 'update' ||
+          method === 'mkdir-index'
+        ) {
+          const filepath = `${dir}/${fullpath}`
+          try {
+            if (method !== 'create-index' && method !== 'mkdir-index') {
+              const { object } = await readObject({ fs, gitdir, oid })
+              if (chmod) {
+                // Note: the mode option of fs.write only works when creating files,
+                // not updating them. Since the `fs` plugin doesn't expose `chmod` this
+                // is our only option.
+                await fs.rm(filepath)
               }
-
-              const stats = await fs.lstat(filepath)
-              // We can't trust the executable bit returned by lstat on Windows,
-              // so we need to preserve this value from the TREE.
-              // TODO: Figure out how git handles this internally.
-              if (mode === 0o100755) {
-                stats.mode = 0o755
+              if (mode === 0o100644) {
+                // regular file
+                await fs.write(filepath, object)
+              } else if (mode === 0o100755) {
+                // executable file
+                await fs.write(filepath, object, { mode: 0o777 })
+              } else if (mode === 0o120000) {
+                // symlink
+                await fs.writelink(filepath, object)
+              } else {
+                throw new InternalError(
+                  `Invalid mode 0o${mode.toString(8)} detected in blob ${oid}`
+                )
               }
-              // Submodules are present in the git index but use a unique mode different from trees
-              if (method === 'mkdir-index') {
-                stats.mode = 0o160000
-              }
-              index.insert({
-                filepath: fullpath,
-                stats,
-                oid,
-              })
-              if (onProgress) {
-                await onProgress({
-                  phase: 'Updating workdir',
-                  loaded: ++count,
-                  total,
-                })
-              }
-            } catch (e) {
-              console.log(e)
             }
-          })
-      )
+
+            const stats = await fs.lstat(filepath)
+            // We can't trust the executable bit returned by lstat on Windows,
+            // so we need to preserve this value from the TREE.
+            // TODO: Figure out how git handles this internally.
+            if (mode === 0o100755) {
+              stats.mode = 0o755
+            }
+            // Submodules are present in the git index but use a unique mode different from trees
+            if (method === 'mkdir-index') {
+              stats.mode = 0o160000
+            }
+            index.insert({
+              filepath: fullpath,
+              stats,
+              oid,
+            })
+            if (onProgress) {
+              await onProgress({
+                phase: 'Updating workdir',
+                loaded: ++count,
+                total,
+              })
+            }
+          } catch (e) {
+            console.log(e)
+          }
+        }
+      }
     })
   }
